refactor(authentication): 优化权限血缘关系

+++ mumu-services/mumu-authentication/authentication-infrastructure/src/main/java/baby/mumu/authentication/infrastructure/authority/gatewayimpl/AuthorityGatewayImpl.java
@@ -26,7 +26,6 @@ import baby.mumu.authentication.infrastructure.authority.gatewayimpl.database.da
 import baby.mumu.authentication.infrastructure.authority.gatewayimpl.database.dataobject.AuthorityDo;
 import baby.mumu.authentication.infrastructure.authority.gatewayimpl.redis.AuthorityRedisRepository;
 import baby.mumu.authentication.infrastructure.relations.database.AuthorityPathsDo;
-import baby.mumu.authentication.infrastructure.relations.database.AuthorityPathsDoId;
 import baby.mumu.authentication.infrastructure.relations.database.AuthorityPathsRepository;
 import baby.mumu.basis.annotations.DangerousOperation;
 import baby.mumu.basis.exception.MuMuException;
@@ -34,6 +33,7 @@ import baby.mumu.basis.response.ResponseCode;
 import baby.mumu.extension.ExtensionProperties;
 import baby.mumu.extension.GlobalProperties;
 import baby.mumu.extension.distributed.lock.DistributedLock;
+import baby.mumu.unique.client.api.PrimaryKeyGrpcService;
 import io.micrometer.observation.annotation.Observed;
 import java.time.Instant;
 import java.util.List;
@@ -75,6 +75,7 @@ public class AuthorityGatewayImpl implements AuthorityGateway {
   private final ExtensionProperties extensionProperties;
   private final AuthorityRedisRepository authorityRedisRepository;
   private final AuthorityPathsRepository authorityPathsRepository;
+  private final PrimaryKeyGrpcService primaryKeyGrpcService;
 
   @Autowired
   public AuthorityGatewayImpl(AuthorityRepository authorityRepository,
@@ -82,7 +83,8 @@ public class AuthorityGatewayImpl implements AuthorityGateway {
     AuthorityConvertor authorityConvertor,
     AuthorityArchivedRepository authorityArchivedRepository, JobScheduler jobScheduler,
     ExtensionProperties extensionProperties, AuthorityRedisRepository authorityRedisRepository,
-    AuthorityPathsRepository authorityPathsRepository) {
+    AuthorityPathsRepository authorityPathsRepository,
+    PrimaryKeyGrpcService primaryKeyGrpcService) {
     this.authorityRepository = authorityRepository;
     this.roleGateway = roleGateway;
     this.authorityConvertor = authorityConvertor;
@@ -92,6 +94,7 @@ public class AuthorityGatewayImpl implements AuthorityGateway {
     this.extensionProperties = extensionProperties;
     this.authorityRedisRepository = authorityRedisRepository;
     this.authorityPathsRepository = authorityPathsRepository;
+    this.primaryKeyGrpcService = primaryKeyGrpcService;
   }
 
   @Override
@@ -106,7 +109,7 @@ public class AuthorityGatewayImpl implements AuthorityGateway {
       .ifPresentOrElse(authorityDo -> {
         authorityRepository.persist(authorityDo);
         authorityPathsRepository.persist(
-          new AuthorityPathsDo(new AuthorityPathsDoId(authorityDo.getId(), authorityDo.getId()),
+          new AuthorityPathsDo(primaryKeyGrpcService.snowflake(),
             authorityDo, authorityDo, 0L));
         authorityRedisRepository.deleteById(authorityDo.getId());
       }, () -> {
@@ -272,19 +275,21 @@ public class AuthorityGatewayImpl implements AuthorityGateway {
         ancestorId);
       // 成环检测
       Set<Long> ancestorIds = ancestorAuthorities.stream()
-        .map(authorityPathsDo -> authorityPathsDo.getId().getAncestorId())
+        .map(authorityPathsDo -> authorityPathsDo.getAncestor().getId())
         .collect(
           Collectors.toSet());
       if (ancestorIds.contains(descendantId)) {
         throw new MuMuException(ResponseCode.AUTHORITY_CYCLE);
       }
-      if (authorityPathsRepository.existsById(new AuthorityPathsDoId(ancestorId, descendantId))) {
+      if (authorityPathsRepository.existsPath(descendantId, ancestorId)) {
         throw new MuMuException(ResponseCode.AUTHORITY_PATH_ALREADY_EXISTS);
       }
       List<AuthorityPathsDo> authorityPathsDos = ancestorAuthorities.stream()
-        .map(ancestorAuthority -> new AuthorityPathsDo(
-          new AuthorityPathsDoId(ancestorAuthority.getId().getAncestorId(), descendantId),
+        .map(ancestorAuthority -> new AuthorityPathsDo(primaryKeyGrpcService.snowflake(),
           ancestorAuthority.getAncestor(), descendantAuthorityDo, ancestorAuthority.getDepth() + 1))
+        .filter(authorityPathsDo -> !authorityPathsRepository.existsPath(
+          authorityPathsDo.getDescendant().getId(), authorityPathsDo.getAncestor()
+            .getId(), authorityPathsDo.getDepth()))
         .collect(
           Collectors.toList());
       authorityPathsRepository.persistAll(authorityPathsDos);
@@ -299,7 +304,7 @@ public class AuthorityGatewayImpl implements AuthorityGateway {
     Page<AuthorityPathsDo> repositoryAll = authorityPathsRepository.findRootAuthorities(
       pageRequest);
     List<Authority> authorities = repositoryAll.getContent().stream().flatMap(
-        authorityPathsDo -> findById(authorityPathsDo.getId().getAncestorId()).stream())
+        authorityPathsDo -> findById(authorityPathsDo.getAncestor().getId()).stream())
       .collect(Collectors.toList());
     return new PageImpl<>(authorities, pageRequest, repositoryAll.getTotalElements());
   }
@@ -311,7 +316,7 @@ public class AuthorityGatewayImpl implements AuthorityGateway {
       ancestorId,
       pageRequest);
     List<Authority> authorities = repositoryAll.getContent().stream().flatMap(
-        authorityPathsDo -> findById(authorityPathsDo.getId().getDescendantId()).stream())
+        authorityPathsDo -> findById(authorityPathsDo.getDescendant().getId()).stream())
       .collect(Collectors.toList());
     return new PageImpl<>(authorities, pageRequest, repositoryAll.getTotalElements());
   }
@@ -319,7 +324,11 @@ public class AuthorityGatewayImpl implements AuthorityGateway {
   @Override
   @Transactional(rollbackFor = Exception.class)
   public void deletePath(Long descendantId, Long ancestorId) {
-    authorityPathsRepository.deleteByDescendantIdAndAncestorId(descendantId, ancestorId);
+    if (authorityPathsRepository.existsDescendantAuthorities(descendantId)) {
+      throw new MuMuException(ResponseCode.DESCENDANT_AUTHORITY_HAS_DESCENDANT_AUTHORITY);
+    }
+    authorityPathsRepository.deleteDirectly(descendantId, ancestorId);
+    authorityPathsRepository.deleteUnreachableData();
     authorityRedisRepository.deleteById(ancestorId);
     authorityRedisRepository.deleteById(descendantId);
   }
